requirements.txt
matplotlib >= 3.6.2
numpy >= 1.23.5
optuna >= 3.0.5
networkx >= 2.8.8
pandas >= 1.5.2
scikit-learn >= 1.2.0
scipy >= 1.9.3
torch_geometric >= 2.2
torchmetrics >= 0.11.0
tqdm >= 4.64.1
PyYAML >= 5.2.6
plotly >= 5.14.1
